refactor: centralize convert_path
