+++ doc_ai/converter/__init__.py
@@ -1,8 +1,10 @@
 from .document_converter import OutputFormat, convert_file, convert_files, suffix_for_format
+from .path import convert_path
 
 __all__ = [
     "OutputFormat",
     "convert_file",
     "convert_files",
     "suffix_for_format",
+    "convert_path",
 ]
