+++ scripts/convert.py
@@ -8,108 +8,9 @@ from dotenv import load_dotenv
 # Allow running without installing as a package by adding project root to path
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from doc_ai import OutputFormat, convert_files, suffix_for_format
-from docling.exceptions import ConversionError
-from doc_ai.metadata import (
-    compute_hash,
-    is_step_done,
-    load_metadata,
-    mark_step,
-    save_metadata,
-)
-
-# File suffixes supported by Docling's ``DocumentConverter``.
-# Anything not in this list will be skipped instead of raising an error when
-# walking directories of mixed content.
-SUPPORTED_SUFFIXES = {
-    ".docx",
-    ".pptx",
-    ".html",
-    ".htm",
-    ".pdf",
-    ".asciidoc",
-    ".adoc",
-    ".md",
-    ".markdown",
-    ".csv",
-    ".xlsx",
-    ".xml",
-    ".json",
-    # Common image formats
-    ".png",
-    ".jpg",
-    ".jpeg",
-    ".gif",
-    ".tif",
-    ".tiff",
-    ".bmp",
-    ".webp",
-    ".svg",
-    # Common audio formats
-    ".wav",
-    ".mp3",
-    ".flac",
-    ".m4a",
-    ".ogg",
-}
+from doc_ai.converter import OutputFormat, convert_path
 
 load_dotenv()
-
-
-def _suffix(fmt: OutputFormat) -> str:
-    """Return desired suffix for ``fmt``."""
-
-    suf = suffix_for_format(fmt)
-    return f".converted{suf}"
-
-
-def convert_path(source: Path, formats: list[OutputFormat]) -> None:
-    """Convert a file or all files under a directory in-place."""
-
-    output_suffixes = {_suffix(fmt) for fmt in OutputFormat}
-
-    def is_output_file(path: Path) -> bool:
-        name = path.name.lower()
-        return any(name.endswith(suf) for suf in output_suffixes)
-
-    def handle_file(file: Path) -> None:
-        """Convert ``file`` if it's not already a derived output and hasn't been processed."""
-
-        if is_output_file(file):
-            return
-        if file.suffix.lower() not in SUPPORTED_SUFFIXES:
-            return
-
-        meta = load_metadata(file)
-        file_hash = compute_hash(file)
-        if meta.blake2b == file_hash and is_step_done(meta, "conversion"):
-            return
-        if meta.blake2b != file_hash:
-            meta.blake2b = file_hash
-            meta.extra = {}
-
-        outputs = {
-            fmt: file.with_name(file.name + _suffix(fmt))
-            for fmt in formats
-            if not (fmt == OutputFormat.MARKDOWN and file.suffix.lower() == ".md")
-        }
-        if not outputs:
-            mark_step(meta, "conversion")
-            save_metadata(file, meta)
-            return
-        try:
-            convert_files(file, outputs)
-        except ConversionError:
-            return
-        mark_step(meta, "conversion")
-        save_metadata(file, meta)
-
-    if source.is_file():
-        handle_file(source)
-    else:
-        for file in source.rglob("*"):
-            if file.is_file():
-                handle_file(file)
 
 
 if __name__ == "__main__":
@@ -119,8 +20,10 @@ if __name__ == "__main__":
         "--format",
         dest="formats",
         action="append",
-        help="Desired output format(s). Can be passed multiple times.\n"
-        "If omitted, the OUTPUT_FORMATS environment variable or Markdown is used.",
+        help=(
+            "Desired output format(s). Can be passed multiple times.\n"
+            "If omitted, the OUTPUT_FORMATS environment variable or Markdown is used."
+        ),
     )
     args = parser.parse_args()
 
@@ -133,7 +36,9 @@ if __name__ == "__main__":
                 formats.append(OutputFormat(val.strip()))
             except ValueError as exc:  # provide clearer error message
                 valid = ", ".join(f.value for f in OutputFormat)
-                raise SystemExit(f"Invalid output format '{val}'. Choose from: {valid}") from exc
+                raise SystemExit(
+                    f"Invalid output format '{val}'. Choose from: {valid}"
+                ) from exc
         return formats
 
     if args.formats:
